requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
pydantic==2.10.6
pydantic-settings==2.7.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
httpx==0.28.1
websockets==14.1
python-dotenv==1.0.1
anthropic>=0.40.0
openai>=1.0.0
cryptography==44.0.0
redis>=5.0.0
jinja2>=3.1.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp-proto-grpc>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-sqlalchemy>=0.46b0
opentelemetry-instrumentation-redis>=0.46b0
opentelemetry-instrumentation-httpx>=0.46b0
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
